restructured files into account folder
