+++ src/index.js
@@ -7,7 +7,7 @@ import { Router, Route, browserHistory } from 'react-router'
 import Landing from './components/landing/Landing'
 import Dashboard from './components/dashboard/Dashboard'
 import reducer from './reducers'
-import AccountDetailsPage from './components/AccountDetailsPage'
+import AccountDetailsPage from './components/account/details/AccountDetailsPage'
 
 import './styles/index.css'
